requirements.txt
# Core dependencies for Twitter Sentiment Analysis
# Author: Lakshya Khetan (lakshyaketan00@gmail.com)

# Core data processing
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Machine Learning & Deep Learning
scikit-learn>=1.1.0
tensorflow>=2.10.0
keras>=2.10.0

# Natural Language Processing
nltk>=3.8
textblob>=0.17.1
tweepy>=4.14.0

# Data Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
wordcloud>=1.9.0
plotly>=5.11.0

# Configuration Management
PyYAML>=6.0
python-dotenv>=0.19.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Jupyter environment
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.15.0
jupyterlab>=3.4.0

# Text preprocessing
preprocessor>=0.6.0
emoji>=2.0.0

# Utility libraries
tqdm>=4.64.0
requests>=2.28.0
pathlib2>=2.3.0

# Optional: Database support
# sqlalchemy>=1.4.0
# pymongo>=4.0.0

# Optional: For faster computation with GPU
# tensorflow-gpu>=2.10.0

# Optional: Advanced NLP models
# transformers>=4.21.0
# torch>=1.12.0